RF: moved Debian related pieces under distributions.debian
