+++ niceman/distributions/base.py
@@ -8,11 +8,17 @@
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Orchestrators helping with management of target environments (remote or local)"""
 
-from importlib import import_module
+import os
 import abc
+
+from importlib import import_module
+from six import viewvalues
+
+from niceman.cmd import Runner
 
 import logging
 lgr = logging.getLogger('niceman.distributions')
+
 
 class Distribution(object):
     """
@@ -79,3 +85,109 @@ class Distribution(object):
             Environment sub-class instance.
         """
         return
+
+# Note: The following was derived from ReproZip's PkgManager class
+# (Revised BSD License)
+
+class PackageTracer(object):
+    """Base class for package trackers.
+
+    ATM :term:`Package` describes all of possible "archives" which deliver
+    some piece of software or data -- packages by distributions (Debian, conda,
+    pip, ...), VCS repositories or even container images -- something which has
+    to be installed to fulfill environment spec
+    """
+
+    def __init__(self, environ=None):
+        # will be (re)used to run external commands, and let's hardcode LC_ALL
+        # codepage just in case since we might want to comprehend error
+        # messages
+        self._environ = environ or Runner(env={'LC_ALL': 'C'})
+
+    def identify_packages_from_files(self, files):
+        """Identifies packages for a given collection of files
+
+        From an iterative collection of files, we identify the packages
+        that contain the files and any files that are not related.
+
+        Parameters
+        ----------
+        files : iterable
+            Container (e.g. list or set) of file paths
+
+        Return
+        ------
+        (found_packages, unknown_files)
+            - found_packages is a list of dicts that holds information about
+              the found packages. Package dicts need at least "name" and
+              "files" (that contains an array of related files)
+            - unknown_files is a list of files that were not found in
+              a package
+        """
+        unknown_files = set()
+        found_packages = {}
+        nb_pkg_files = 0
+
+        file_to_package_dict = self._get_packagenames_for_files(files)
+        for f in files:
+            if not os.path.lexists(f):
+                lgr.warning(
+                    "Provided file %s doesn't exist, spec might be incomplete",
+                    f)
+            # Stores the file
+            if f not in file_to_package_dict:
+                unknown_files.add(f)
+            else:
+                # TODO: pkgname should become  pkgid
+                # where for packages from distributions would be name,
+                # for VCS -- their path
+                pkgname = file_to_package_dict[f]
+                if pkgname is None:
+                    unknown_files.add(f)
+                elif pkgname in found_packages:
+                    found_packages[pkgname]["files"].append(f)
+                    nb_pkg_files += 1
+                else:
+                    pkg = self._create_package(pkgname)
+                    if pkg:
+                        found_packages[pkgname] = pkg
+                        pkg["files"].append(f)
+                        nb_pkg_files += 1
+                    else:
+                        unknown_files.add(f)
+
+        lgr.info("%s: %d packages with %d files, and %d other files",
+                 self.__class__.__name__,
+                 len(found_packages),
+                 nb_pkg_files,
+                 len(unknown_files))
+
+        return list(viewvalues(found_packages)), list(unknown_files)
+
+    def identify_package_origins(self, packages):
+        """Identify and collate origins from a set of packages
+
+        From a collection of packages, identify the unique origins
+        into a separate collection.
+
+        Parameters
+        ----------
+        packages : iterable
+            Array of Package (to be updated)
+
+        Return
+        ------
+        (origins)
+            - Discovered collection of origins
+        """
+        raise NotImplementedError
+
+    def _get_packagenames_for_files(self, files):
+        raise NotImplementedError
+
+    def _create_package(self, pkgname):
+        """Creates implementation specific Package object
+
+        (well -- atm still an OrderedDict)
+        """
+        raise NotImplementedError
